+++ frontend/src/metabase/common/components/DatePicker/SpecificDatePicker/DateRangePicker/DateRangePickerBody/index.ts
@@ -1,0 +1,1 @@
+export * from "./DateRangePickerBody";
