+++ frontend/src/metabase/common/components/DatePicker/SpecificDatePicker/DateRangePicker/SimpleDateRangePicker/index.ts
@@ -1,2 +1,1 @@
-export * from "./DateRangePicker";
 export * from "./SimpleDateRangePicker";
