+++ frontend/src/metabase/common/components/DatePicker/SpecificDatePicker/SimpleSpecificDatePicker/index.ts
@@ -1,2 +1,1 @@
-export * from "./SpecificDatePicker";
 export * from "./SimpleSpecificDatePicker";
