+++ frontend/src/metabase/common/components/DatePicker/SpecificDatePicker/SingleDatePicker/index.ts
@@ -1,1 +1,2 @@
 export * from "./SingleDatePicker";
+export * from "./SimpleSingleDatePicker";
